+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,12 +1,9 @@
-package chess.domain;
+package chess.domain.board;
 
-import chess.domain.piece.Bishop;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.*;
+import chess.domain.square.File;
+import chess.domain.square.Rank;
+import chess.domain.square.Square;
 
 import java.util.Arrays;
 import java.util.HashMap;
