+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,8 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.PieceColor;
-import chess.domain.PieceType;
-import chess.domain.Square;
+import chess.domain.square.Square;
 
 public abstract class Piece {
