+++ src/main/java/chess/domain/square/Rank.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.square;
 
 import java.util.Arrays;
 import java.util.regex.Pattern;
