+++ src/test/java/chess/domain/board/BoardTest.java
@@ -1,9 +1,12 @@
-package chess.domain;
+package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import chess.domain.board.Board;
+import chess.domain.piece.PieceColor;
 import chess.domain.piece.Rook;
+import chess.domain.square.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
