+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -2,8 +2,7 @@ package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.PieceColor;
-import chess.domain.Square;
+import chess.domain.square.Square;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
