+++ src/test/java/chess/domain/square/RankTest.java
@@ -1,7 +1,8 @@
-package chess.domain;
+package chess.domain.square;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import chess.domain.square.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
